refactor: split events.js into focused modules
Split the 714-line events.js into 4 focused modules for better maintainability:

- textarea-utils.js (217 lines): Text manipulation utilities
  - Line position/column calculation
  - Indent/unindent operations
  - Selection handling

- list-editor.js (290 lines): Markdown list operations
  - Tab/Backspace handling for lists
  - Ordered list renumbering
  - Task list continuation

- auto-save.js (38 lines): Auto-save logic
  - Debounced save scheduling
  - Status updates

- events.js (355 lines): Event handlers only
  - All event listener attachments
  - High-level event handlers

Total reduction: events.js 714 → 355 lines (50% smaller)
All modules use ES Modules (import/export) for clean dependency management.
